chore: rename keyboard
